+++ examples/simple/config/config.exs
@@ -1,6 +1,6 @@
 use Mix.Config
 
-config :simple, Repo,
+config :simple, Simple.Repo,
   database: "ecto_simple",
   username: "postgres",
   password: "postgres",