+++ lib/ecto/migration.ex
@@ -33,10 +33,10 @@ defmodule Ecto.Migration do
 
   Ecto provides some mix tasks to help developers work with migrations:
 
-    * `mix ecto.gen.migration Repo add_weather_table` - generates a
+    * `mix ecto.gen.migration add_weather_table` - generates a
       migration that the user can fill in with particular commands
-    * `mix ecto.migrate Repo` - migrates a repository
-    * `mix ecto.rollback Repo` - rolls back a particular migration
+    * `mix ecto.migrate` - migrates a repository
+    * `mix ecto.rollback` - rolls back a particular migration
 
   Run the `mix help COMMAND` for more information.