+++ lib/mix/tasks/ecto.create.ex
@@ -2,18 +2,20 @@ defmodule Mix.Tasks.Ecto.Create do
   use Mix.Task
   import Mix.Ecto
 
-  @shortdoc "Create the database for the repo"
+  @shortdoc "Create the storage for the repo"
 
   @moduledoc """
-  Create the database for the given repository, as specified in the repo's `url`.
-
-  ## Command line options
-
-  * `--no-start` - do not start applications
+  Create the storage for the repository.
 
   ## Examples
 
-      mix ecto.create MyApp.Repo
+      mix ecto.create
+      mix ecto.create -r Custom.Repo
+
+  ## Command line options
+
+    * `-r`, `--repo` - the repo to create (defaults to `YourApp.Repo`)
+    * `--no-start` - do not start applications
 
   """
 
@@ -21,7 +23,7 @@ defmodule Mix.Tasks.Ecto.Create do
   def run(args) do
     Mix.Task.run "app.start", args
 
-    {repo, _} = parse_repo(args)
+    repo = parse_repo(args)
     ensure_repo(repo)
     ensure_implements(repo.adapter, Ecto.Adapter.Storage, "to create storage for #{inspect repo}")
 
@@ -31,8 +33,7 @@ defmodule Mix.Tasks.Ecto.Create do
       {:error, :already_up} ->
         Mix.shell.info "The database for repo #{inspect repo} has already been created."
       {:error, term} ->
-        raise Mix.Error, message:
-           "The database for repo #{inspect repo} couldn't be created, reason given: #{term}."
+        Mix.raise "The database for repo #{inspect repo} couldn't be created, reason given: #{term}."
     end
   end
 end