+++ lib/mix/tasks/ecto.migrate.ex
@@ -15,36 +15,37 @@ defmodule Mix.Tasks.Ecto.Migrate do
   to a version number, supply `--to version_number`.
   To migrate up a specific number of times, use `--step n`.
 
-  ## Command line options
-
-  * `--all` - run all pending migrations
-  * `--step` / `-n` - run n number of pending migrations
-  * `--to` / `-v` - run all migrations up to and including version
-  * `--no-start` - do not start applications
-
   ## Examples
 
-      mix ecto.migrate MyApp.Repo
+      mix ecto.migrate
+      mix ecto.migrate -r Custom.Repo
 
-      mix ecto.migrate MyApp.Repo -n 3
-      mix ecto.migrate MyApp.Repo --step 3
+      mix ecto.migrate -n 3
+      mix ecto.migrate --step 3
 
-      mix ecto.migrate MyApp.Repo -v 20080906120000
-      mix ecto.migrate MyApp.Repo --to 20080906120000
+      mix ecto.migrate -v 20080906120000
+      mix ecto.migrate --to 20080906120000
 
+  ## Command line options
+
+    * `-r`, `--repo` - the repo to migrate (defaults to `YourApp.Repo`)
+    * `--all` - run all pending migrations
+    * `--step` / `-n` - run n number of pending migrations
+    * `--to` / `-v` - run all migrations up to and including version
+    * `--no-start` - do not start applications
   """
 
   @doc false
   def run(args, migrator \\ &Ecto.Migrator.run/4) do
     Mix.Task.run "app.start", args
+    repo = parse_repo(args)
 
-    {opts, args, _} = OptionParser.parse args,
+    {opts, _, _} = OptionParser.parse args,
       switches: [all: :boolean, step: :integer, to: :integer, start: :boolean],
       aliases: [n: :step, v: :to]
 
-    {repo, _} = parse_repo(args)
     ensure_repo(repo)
-    unless opts[:start], do: ensure_started(repo)
+    if Keyword.get(opts, :start, true), do: ensure_started(repo)
 
     unless opts[:to] || opts[:step] || opts[:all] do
       opts = Keyword.put(opts, :all, true)