+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * `Ecto.Adapters.Postgres.query/4` has been renamed to `Ecto.Adapters.SQL.query/4`
   * `Ecto.Adapters.Postgres.begin_test_transaction/2` has been renamed to `Ecto.Adapters.SQL.begin_test_transaction/2`
   * `Ecto.Adapters.Postgres.rollback_test_transaction/2` has been renamed to `Ecto.Adapters.SQL.rollback_test_transaction/2`
+  * Mix tasks now expect the repository with the option `-r`, otherwise it defaults to the application based one 
 
 ## v0.6.0 (2015-01-17)